Move line and gutter handling to the Frame struct
The Frame struct now contains the code to create horizontal lines and
the left-hand gutter (numbered and unnumbered).  It remembers the width
of the last number printed, so that unnumbered gutters and the closing
horizontal line are spaced consistently. (There's an unavoidable break
in the vertical line to the right of the line numbers between 9999 and
10000. Hopefully there will be very few files with over ten thousand
lines!)  Here's an example:

─────┬──────────────
9999 │
     │
10000 │
      │
──────┴─────────────