Show non-printables after syntax coloring.
Turning control characters like tabs and newlines into visible glyphs on
the entire line, before syntax highlighting, can break highlighting.
Instead, we do it on a region-by-region basis, so the highlighting for
keywords and variables doesn't get confused by the character not in the
language.

That makes the stored results change again.