Add cursor_max to Frame
We were using UnicodeWidthStr to calculate the width of the gutter.
We'll need that when we're processing Unicode graphemes, but we don't
need it yet.  And we were using the .len() of " │ ", which is 5, instead
of it's width in columns (3).  There's now a test for that in case we
regress, and the cursor_max() method on Frame gives the number of
columns left after the gutter is printed for the term_width given to the
Frame.